requirements.txt
Django==3.1.7
psycopg2-binary
python-dotenv~=0.12.0
requests~=2.23.0
django-model-utils~=4.0.0
pytest~=5.3.5
pytest-django
pytest-xdist
sentry-sdk==0.14.4
djangorestframework
djangorestframework-recursive~=0.1.2
python-logstash
pyfunctional
channels~=2.4.0
coreschema~=0.0.4
django-cors-headers
django-datetime-helpers
coreapi~=2.3.3
djangorestframework_simplejwt
dramatiq[rabbitmq,watch]
periodiq
django_dramatiq
git+https://github.com/eugen1j/django_periodiq
sentry-dramatiq
slack_sdk
Pillow
django-cleanup~=5.1.0
